graphql: Add search query openAt filter for venues
Implemented search query openAt filter for venues, which limits the
results to those venues that are open at the given time.

The filter accepts ISO 8601 datetimes with and without a timezone
offset (if there is no offset, the datetime is assumed to be in
timezone "Europe/Helsinki") and values conforming to Elastic Search
date math, like "now+3h".