Remove Linked Events dependency from location import
Before this changem every imported location was enriched with place data
from Linked Events. The enriched data wasn't used for much, and doing
the enrichment caused a lot of requests to LE making the location import
slow and less robust, so it made sense to get rid of that.

The only thing that was actually used from the enriched data were
administrative divisions. For those there is now a separate importer
which uses django-munigeo for fetching the divisions. Because munigeo
requires a database and US isn't actually using one ATM, a temporary
Spatialite database is used just for storing the divisions for the
duration of the imports.