feat(sources): add LocationImporterAPI to clean and reusability
LIIKUNTA-578.

Add a LocationImporterAPI to make it easier or even possible to patch
the return values of the mocked external services.

Also, now all the external services are listed in a one class (the LocationImporterAPI)
and the utilities that uses it are separated from the actual data fetching.

Move geo import of Helsinki divisions to new function: Make the call of `call_command("geo_import", "helsinki", "--divisions")` mockable.

Wrote some missing tests for the location importer and the new reservable-field.

Ran the code formatting tools.