feat(graphql): exact matches should be higher in relevance
US-123
Use bool-query with `match_bool_prefix` and `match_phrase` instead of `query_string` query.

Add a bool-query to replace the query_string-query,
since it is simpler and does not need any special care for input.

Add a `match_phrase` query on side of `match_bool_prefix`
to give higher score for "exact matches".

Add a new query function on besides of the old 
`query_string` implementation, just to make the future development easier, if we go back to query string.

Example (tested): When searching with a text-parameter set as 
"Itäkeskuksen uimahalli / Kuntosali",  it returns 
the exact match as first. The result set is a following 
(only name fields included):
1. Itäkeskuksen uimahalli / Kuntosali
2. Itäkeskuksen uimahalli
3. Itäkeskuksen uimahalli / Liikuntasali
4. Itäkeskuksen palvelutalo