feat(graphql): upgrade to eslint v9 + add minimal types to src/index.ts
Previously graphql/src/index.ts had a lot of `any` type usage.

Add minimal types to make at least part of the expected data structure
more apparent. The types are very sparse still, and in the spirit of
DRY it'd be good to e.g. automatically generate the TypeScript types
from the GraphQL / gql types already present in the codebase, if
possible.

refs US-165