fix(sources): remove incorrect deprecation
Previously get_unit_id_to_accessibility_sentences_mapping was
deprecated based on the idea that accessibility sentences for a unit can
be fetched using LocationImporterAPI.fetch_tpr_units function.

Looking at that function it fetches data from
https://www.hel.fi/palvelukarttaws/rest/v4/unit/?newfeatures=yes
which DOES NOT include accessibility sentences.

The deprecation text referred to a single unit's fetching e.g.
https://www.hel.fi/palvelukarttaws/rest/v4/unit/42284?official=yes&format=json&newfeatures=yes
which DOES include accessibility sentences for that unit but it is not
the URL that is being used by LocationImporterAPI.fetch_tpr_units.
Thus, the deprecation was based on incorrect information.

Removed the deprecation as it was based on incorrect information.

refs US-174