feat(graphql): use better workaround for readiness build time
Similarly as in sources app's settings.py a file's modification time is
used as a workaround for approximating readiness endpoint's build time.
This is not foolproof, but better than nothing—the previous way of using
the file's evaluation time was probably more problematic than this as it
probably would've changed if the pod got respawned.

refs US-151