fix(graphql)!: remove needless recursion from LocationDescription type
NOTE: This commit has been marked as BREAKING CHANGE because of removing
      an existing field from the API.

LocationDescription → UnifiedSearchVenue → LocationDescription → ...
ad infinitum i.e. LocationDescription was infinitely recursive structure
...venue.location.venue.location.venue.location... without any known
need for this recursion.

Remove the needless recursion to simplify the unifiedSearch query and
limit its maximum depth.

I did take a look at events-helsinki-monorepo, which is the most
extensive user of unified search AFAIK, and it didn't seem to be using
this recursion.

refs US-181