feat(sources): remove support for ingest_data data deletion as unneeded
The support for deleting data using `ingest_data` management command
was broken (specifically location importer didn't just delete data when
asked to but it ALSO imported data even when it shouldn't have), and is
not used in Azure DevOps, so it can be safely removed.

This is simplifying the codebase by removing broken functionality that
"might be useful" in a sense, but probably isn't *really* needed.

refs US-185