feat(graphql): allow apollo sandbox in production
Rationale:

Had a talk with the team about this change today in daily, the main
argument why apollo sandbox can be enabled in production is that the
data available through the GraphQL search API is not sensitive i.e. it
is open and public data. There is no authorization nor authentication
used in the API, so there are no concerns about exposing private data
through it.

refs US-181